+++ src/main/kotlin/io/projectenv/intellijplugin/configurers/gradle/GradleConfigurer.kt
@@ -1,10 +1,12 @@
 package io.projectenv.intellijplugin.configurers.gradle
 
+import com.intellij.openapi.components.service
 import com.intellij.openapi.externalSystem.settings.ExternalSystemSettingsListenerAdapter
 import com.intellij.openapi.project.Project
 import io.projectenv.core.tools.info.GradleInfo
 import io.projectenv.core.tools.info.ToolInfo
 import io.projectenv.intellijplugin.configurers.ToolConfigurer
+import io.projectenv.intellijplugin.services.ProjectEnvService
 import org.jetbrains.plugins.gradle.settings.DistributionType
 import org.jetbrains.plugins.gradle.settings.GradleProjectSettings
 import org.jetbrains.plugins.gradle.settings.GradleSettings
@@ -15,7 +17,8 @@ class GradleConfigurer(project: Project) : ToolConfigurer<GradleInfo>,
     private var toolInfo: ToolInfo? = null
 
     init {
-        GradleSettings.getInstance(project).subscribe(this, this)
+        val disposableParent = project.service<ProjectEnvService>()
+        GradleSettings.getInstance(project).subscribe(this, disposableParent)
     }
 
     override fun supportsType(toolInfo: ToolInfo): Boolean {
@@ -27,17 +30,14 @@ class GradleConfigurer(project: Project) : ToolConfigurer<GradleInfo>,
     }
 
     override fun onProjectsLinked(settings: MutableCollection<GradleProjectSettings>) {
-        if (toolInfo == null) {
-            return
-        }
-
-        for (projectSettings in settings) {
-            projectSettings.distributionType = DistributionType.LOCAL
-            projectSettings.gradleHome = toolInfo!!.location.canonicalPath
-        }
+        updateProjectsConfiguration(settings)
     }
 
     override fun onProjectsLoaded(settings: MutableCollection<GradleProjectSettings>) {
+        updateProjectsConfiguration(settings)
+    }
+
+    private fun updateProjectsConfiguration(settings: MutableCollection<GradleProjectSettings>) {
         if (toolInfo == null) {
             return
         }