+++ src/main/kotlin/io/projectenv/intellijplugin/services/ProjectEnvService.kt
@@ -1,6 +1,8 @@
 package io.projectenv.intellijplugin.services
 
-interface ProjectEnvService {
+import com.intellij.openapi.Disposable
+
+interface ProjectEnvService : Disposable {
 
     fun refreshProjectEnv()